+++ server/internal/model/entity/admin_user.go
@@ -26,6 +26,6 @@ type AdminUser struct {
 	LastLoginIp string `json:"lastLoginIp" orm:"last_login_ip" description:"最后登录IP"`          // 最后登录IP
 	CreatedBy   uint64 `json:"createdBy"   orm:"created_by"    description:"创建人ID"`           // 创建人ID
 	UpdatedBy   uint64 `json:"updatedBy"   orm:"updated_by"    description:"更新人ID"`           // 更新人ID
-	CreateTime  uint64 `json:"createTime"  orm:"create_time"   description:"创建时间"`            // 创建时间
-	UpdateTime  uint64 `json:"updateTime"  orm:"update_time"   description:"更新时间"`            // 更新时间
+	CreateTime  uint   `json:"createTime"  orm:"create_time"   description:"创建时间"`            // 创建时间
+	UpdateTime  uint   `json:"updateTime"  orm:"update_time"   description:"更新时间"`            // 更新时间
 }
