+++ server/internal/model/entity/captcha.go
@@ -10,6 +10,6 @@ type Captcha struct {
 	Key        string `json:"key"        orm:"key"         description:"验证码key（MD5）"`         // 验证码key（MD5）
 	Code       string `json:"code"       orm:"code"        description:"验证码code（MD5）"`        // 验证码code（MD5）
 	Captcha    string `json:"captcha"    orm:"captcha"     description:"验证码数据（JSON，包含点位坐标等）"` // 验证码数据（JSON，包含点位坐标等）
-	CreateTime uint64 `json:"createTime" orm:"create_time" description:"创建时间（unix秒）"`         // 创建时间（unix秒）
-	ExpireTime uint64 `json:"expireTime" orm:"expire_time" description:"过期时间（unix秒）"`         // 过期时间（unix秒）
+	CreateTime uint   `json:"createTime" orm:"create_time" description:"创建时间（unix秒）"`         // 创建时间（unix秒）
+	ExpireTime uint   `json:"expireTime" orm:"expire_time" description:"过期时间（unix秒）"`         // 过期时间（unix秒）
 }
